Add files via upload
Hey I found some info online on how to make tables in the GUI, so this is some code I found and I modified it to what I think the PCB table should look like. Obviously more needs to be changed and refactored but I think I got the main idea.