Add files via upload
need to change file name to match class (or vice versa) 
This is the setup for the command line in our gui. I still need to figure out how to icorporate it with the table.
Also needs to be changed and refactored.
Sample text is outputted for the proc(), mem(), load(), exe(), and reset() functions if you type them in. I hope this helps.